requirements.txt
asarPy==1.0.1
beautifulsoup4==4.13.5
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
Flask==3.1.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
MouseInfo==0.1.3
pillow==11.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.1.1
python3-xlib==0.15
pytweening==1.2.0
pywhatkit==5.4
requests==2.32.5
soupsieve==2.7
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wikipedia==1.4.0
